requirements.txt
requests==2.31.0
transformers==4.36.0
python-dotenv==1.2.1
openai>=1.0.0
